+++ src/main/java/com/petfam/petfam/service/like/LikeService.java
@@ -1,8 +1,8 @@
 package com.petfam.petfam.service.like;
 
-import com.petfam.petfam.dto.comment.CommentLikeResponseDto;
-import com.petfam.petfam.dto.post.PostLikeResponseDto;
-import com.petfam.petfam.dto.recomment.ReCommentLikeResponseDto;
+import com.petfam.petfam.dto.like.CommentLikeResponseDto;
+import com.petfam.petfam.dto.like.PostLikeResponseDto;
+import com.petfam.petfam.dto.like.ReCommentLikeResponseDto;
 import com.petfam.petfam.entity.User;
 
 public interface LikeService {
